Test.Cardano.Chain.Byron.API

Documentation

genApplyMempoolPayloadErr ∷ ProtocolMagicId → Gen ApplyMempoolPayloadErr

ts_mempoolValidation ∷ TSProperty

A transaction should validate in the mempool at a given slot when it validates in a block issued for that same slot.

ts_roundTripApplyMempoolPayloadErrCompat ∷ TSProperty

ts_scheduledDelegations ∷ TSProperty

getDelegationMap . applyChainTick slot == previewDelegationMap slot

tests ∷ TSGroup