Test.Cardano.Chain.Elaboration.Block

Description

This module provides functionality for translating abstract blocks into concrete blocks. The abstract blocks are generated according the small-step rules for the blockchain (also called the blockchain specification).

Documentation

abEnvToCfg ∷ Environment CHAIN → Config

Make a genesis configuration from an initial abstract environment of the trace.

elaborate ∷ AbstractToConcreteIdMaps → Config → DCert → ChainValidationState → Block → (Block, AbstractToConcreteIdMaps)

Elaborate an abstract block into a concrete block (without annotations).

elaborateBS ∷ AbstractToConcreteIdMaps → Config → DCert → ChainValidationState → Block → (ABlock ByteString, AbstractToConcreteIdMaps)

rcDCert

Arguments

∷ HasCallStack

⇒ VKey
Key for which the delegation certificate is being constructed.

→ BlockCount
Chain stability parameter

→ State CHAIN

→ DCert

Re-construct an abstract delegation certificate from the abstract state.

We need to do this because the delegation certificate is included in the block.

data AbstractToConcreteIdMaps

Constructors

AbstractToConcreteIdMaps

Fields

transactionIds ∷ !(Map TxId TxId)

proposalIds ∷ !(Map UpId UpId)